vasp计算声子谱支持使用kspacing

+++ vasp/vasp_submitjob.py
@@ -104,7 +104,7 @@ class vasp_submitjob:
             src_poscar = Path(self.work_path).joinpath(poscar_number) ; dst_poscar = Path(dst_number_dir).joinpath("POSCAR"); shutil.copy(src_poscar, dst_poscar)
             src_potcar = Path(self.work_path).joinpath("POTCAR")      ; dst_potcar = Path(dst_number_dir).joinpath("POTCAR"); shutil.copy(src_potcar, dst_potcar)
             src_incar  = Path(self.work_path).joinpath("INCAR")       ; dst_incar  = Path(dst_number_dir).joinpath("INCAR" ); shutil.copy(src_incar, dst_incar )
-            src_kpoints= Path(self.work_path).joinpath("KPOINTS")     ; dst_kpoints= Path(dst_number_dir).joinpath("KPOINTS"); shutil.copy(src_kpoints, dst_kpoints)
+            src_kpoints= Path(self.work_path).joinpath("KPOINTS")     ; dst_kpoints= Path(dst_number_dir).joinpath("KPOINTS"); shutil.copy(src_kpoints, dst_kpoints) if src_kpoints.exists() else None
             src_submit = Path(self.work_path).joinpath(jobname)       ; dst_submit = Path(dst_number_dir).joinpath(jobname);  shutil.copy(src_submit, dst_submit)
             cwd = src_poscar.cwd()
             dst_dir = dst_number_dir.absolute()
